requirements.txt
aiogram
google-search-python
